feature/structure-changges-in-traits: SetterTrait splitted into four traits
