+++ src/Traits/ConfigTrait.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Aosmak\Laravel\Layer\Sdk\Traits;
+
+/**
+ * Trait ConfigTrait
+ * @package namespace Aosmak\Laravel\Layer\Sdk\Traits;
+ */
+trait ConfigTrait
+{
+    /**
+     * Config
+     *
+     * @var array contains configuration variables
+     */
+    protected $config;
+
+    /**
+     * Set config
+     *
+     * @param array $config contains configuration variables
+     *
+     * @return void
+     */
+    public function setConfig(array $config)
+    {
+        $this->config = $config;
+    }
+}
